+++ querymate/__init__.py
@@ -63,10 +63,12 @@ from .core.filter import TruePredicate as TruePredicate
 from .core.grouping import DateGranularity as DateGranularity
 from .core.grouping import GroupByConfig as GroupByConfig
 from .core.grouping import GroupedResponse as GroupedResponse
+from .core.grouping import GroupingOptions as GroupingOptions
 from .core.grouping import GroupKeyExtractor as GroupKeyExtractor
 from .core.grouping import GroupResult as GroupResult
 from .core.query_builder import JoinType as JoinType
 from .core.query_builder import QueryBuilder as QueryBuilder
+from .core.querymate import PaginationOptions as PaginationOptions
 from .core.querymate import Querymate as Querymate
 from .types import PaginatedResponse as PaginatedResponse
 from .types import PaginationInfo as PaginationInfo